Fix Go back-compatibility in Accept error check.
Emit the message as DEBUG, so logs avoid false positive errors.

net.ErrClosed was added in Go 1.16, but this project's go.mod specifies
Go 1.13.  We alias Go's internal error type, to avoid the
worse alternative of comparing the string of the error message.

Go maintainers discuss this topic at length in
 https://github.com/golang/go/issues/4373.
Occurences of "use of closed network connection"
are not an actual problem.  It's advisory only, akin to io.EOF.